package-indexer: always calculate md5 locally
